Diabetic_Retinopathy_Detection/PyTorch/requirements.txt
torch==1.9.0
torchvision==0.10.0
Pillow==8.3.2
opencv-python==4.5.3.56
numpy==1.19.5
pandas==1.3.3
scikit-learn==0.24.2
Flask==2.1.1
